Feat: Moved preferred currency from user to project in the backend
