+++ EasyFinance.Application/DTOs/AccessControl/UserRequestDTO.cs
@@ -6,7 +6,6 @@ namespace EasyFinance.Application.DTOs.AccessControl
     {
         public string FirstName { get; set; } = "Default";
         public string LastName { get; set; } = "Default";
-        public string PreferredCurrency { get; set; } = string.Empty;
         public Guid DefaultProject {  get; set; } = Guid.Empty;
     }
 }
