+++ EasyFinance.Application/DTOs/AccessControl/UserResponseDTO.cs
@@ -13,7 +13,6 @@ namespace EasyFinance.Application.DTOs.AccessControl
                 Email = user.Email;
                 FirstName = user.FirstName;
                 LastName = user.LastName;
-                PreferredCurrency = user.PreferredCurrency;
                 Enabled = user.Enabled;
                 IsFirstLogin = user.HasIncompletedInformation;
                 EmailConfirmed = user.EmailConfirmed;
@@ -26,7 +25,6 @@ namespace EasyFinance.Application.DTOs.AccessControl
         public string Email { get; set; } = string.Empty;
         public string FirstName { get; set; } = string.Empty;
         public string LastName { get; set; } = string.Empty;
-        public string PreferredCurrency { get; set; } = string.Empty;
         public bool Enabled { get; set; }
         public bool IsFirstLogin { get; set; }
         public bool EmailConfirmed { get; set; }
