+++ EasyFinance.Application/DTOs/FinancialProject/ProjectRequestDTO.cs
@@ -5,5 +5,6 @@ namespace EasyFinance.Application.DTOs.FinancialProject
     public class ProjectRequestDTO
     {
         public string Name { get; set; } = string.Empty;
+        public string PreferredCurrency { get; set; } = string.Empty;
     }
 }
