+++ EasyFinance.Application/DTOs/FinancialProject/ProjectResponseDTO.cs
@@ -6,5 +6,6 @@ namespace EasyFinance.Application.DTOs.FinancialProject
     {
         public Guid Id { get; set; }
         public string Name { get; set; } = string.Empty;
+        public string PreferredCurrency { get; set; } = string.Empty;
     }
 }
