+++ EasyFinance.Domain.Tests/FinancialProject/ProjectTests.cs
@@ -1,4 +1,5 @@
-﻿using EasyFinance.Common.Tests.FinancialProject;
+﻿using EasyFinance.Common.Tests.AccessControl;
+using EasyFinance.Common.Tests.FinancialProject;
 using EasyFinance.Infrastructure;
 using EasyFinance.Infrastructure.Exceptions;
 using FluentAssertions;
@@ -37,6 +38,29 @@ namespace EasyFinance.Domain.Tests.FinancialProject
             action.Should().Throw<ValidationException>()
                 .WithMessage(string.Format(ValidationMessages.PropertyCantBeNull, "Incomes"))
                 .And.Property.Should().Be("Incomes");
+        }
+
+        [Theory]
+        [InlineData(null)]
+        [InlineData("")]
+        public void AddPreferredCurrency_SendNullAndEmpty_ShouldThrowException(string preferredCurrency)
+        {
+            var action = () => new ProjectBuilder().AddPreferredCurrency(preferredCurrency).Build();
+
+            action.Should().Throw<ValidationException>()
+                .WithMessage(string.Format(ValidationMessages.PropertyCantBeNullOrEmpty, "PreferredCurrency"))
+                .And.Property.Should().Be("PreferredCurrency");
+        }
+
+        [Theory]
+        [InlineData("Test")]
+        public void AddPreferredCurrency_SendInvalid_ShouldThrowException(string preferredCurrency)
+        {
+            var action = () => new ProjectBuilder().AddPreferredCurrency(preferredCurrency).Build();
+
+            action.Should().Throw<ValidationException>()
+                .WithMessage(ValidationMessages.InvalidCurrencyCode)
+                .And.Property.Should().Be("PreferredCurrency");
         }
     }
 }
