+++ EasyFinance.Domain/AccessControl/User.cs
@@ -2,7 +2,6 @@
 using EasyFinance.Infrastructure;
 using EasyFinance.Infrastructure.DTOs;
 using EasyFinance.Infrastructure.Exceptions;
-using EasyFinance.Infrastructure.Validators;
 using Microsoft.AspNetCore.Identity;
 
 namespace EasyFinance.Domain.AccessControl
@@ -16,18 +15,16 @@ namespace EasyFinance.Domain.AccessControl
             Id = id;
         }
 
-        public User(string firstName = "Default", string lastName = "Default", string preferredCurrency = "EUR", bool enabled = default)
+        public User(string firstName = "Default", string lastName = "Default", bool enabled = default)
         {
             FirstName = firstName;
             LastName = lastName;
-            PreferredCurrency = preferredCurrency;
             Enabled = enabled;
         }
 
         public string FirstName { get; private set; } = string.Empty;
         public string LastName { get; private set; } = string.Empty;
         public string FullName => $"{FirstName} {LastName}";
-        public string PreferredCurrency { get; private set; } = string.Empty;
         public bool Enabled { get; set; } = true;
         public bool HasIncompletedInformation => string.IsNullOrEmpty(FirstName) && string.IsNullOrEmpty(LastName);
         public Guid? DefaultProjectId { get; private set; } = default;
@@ -46,17 +43,6 @@ namespace EasyFinance.Domain.AccessControl
                 throw new ValidationException(nameof(LastName), string.Format(ValidationMessages.PropertyCantBeNullOrEmpty, nameof(LastName)));
 
             LastName = lastName;
-        }
-
-        public void SetPreferredCurrency(string preferredCurrency)
-        {
-            if (string.IsNullOrEmpty(preferredCurrency))
-                throw new ValidationException(nameof(PreferredCurrency), string.Format(ValidationMessages.PropertyCantBeNullOrEmpty, nameof(PreferredCurrency)));
-
-            if (!CurrencyValidator.IsValidCurrencyCode(preferredCurrency))
-                throw new ValidationException(nameof(PreferredCurrency), ValidationMessages.InvalidCurrencyCode);
-
-            PreferredCurrency = preferredCurrency;
         }
 
         public AppResponse SetDefaultProject(Guid? projectId)
