+++ EasyFinance.Domain/FinancialProject/Project.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using EasyFinance.Domain.Financial;
 using EasyFinance.Infrastructure;
 using EasyFinance.Infrastructure.Exceptions;
+using EasyFinance.Infrastructure.Validators;
 
 namespace EasyFinance.Domain.FinancialProject
 {
@@ -10,16 +11,18 @@ namespace EasyFinance.Domain.FinancialProject
     {
         private Project() { }
 
-        public Project(Guid id = default, string name = "default", ICollection<Category> categories = default, ICollection<Income> incomes = default)
+        public Project(Guid id = default, string name = "default", string preferredCurrency = "EUR", ICollection<Category> categories = default, ICollection<Income> incomes = default)
             : base(id)
         {
             SetName(name);
             SetCategories(categories ?? new List<Category>());
+            SetPreferredCurrency(preferredCurrency);
             SetIncomes(incomes ?? new List<Income>());
         }
 
         public string Name { get; private set; } = string.Empty;
         public bool IsArchived { get; private set; }
+        public string PreferredCurrency { get; private set; } = string.Empty;
         public ICollection<Category> Categories { get; private set; } = new List<Category>();
         public ICollection<Income> Incomes { get; private set; } = new List<Income>();
 
@@ -53,6 +56,17 @@ namespace EasyFinance.Domain.FinancialProject
                 throw new ValidationException(nameof(Incomes), string.Format(ValidationMessages.PropertyCantBeNull, nameof(Incomes)));
 
             Incomes = incomes;
+        }
+
+        public void SetPreferredCurrency(string preferredCurrency)
+        {
+            if (string.IsNullOrEmpty(preferredCurrency))
+                throw new ValidationException(nameof(PreferredCurrency), string.Format(ValidationMessages.PropertyCantBeNullOrEmpty, nameof(PreferredCurrency)));
+
+            if (!CurrencyValidator.IsValidCurrencyCode(preferredCurrency))
+                throw new ValidationException(nameof(PreferredCurrency), ValidationMessages.InvalidCurrencyCode);
+
+            PreferredCurrency = preferredCurrency;
         }
 
         public void SetName(string name)
