+++ EasyFinance.Persistence/Mapping/AccessControl/UserConfiguration.cs
@@ -24,9 +24,6 @@ namespace EasyFinance.Persistence.Mapping.AccessControl
             builder.Property(p => p.Enabled)
                 .IsRequired();
 
-            builder.Property(p => p.PreferredCurrency)
-                .HasMaxLength(3);
-
             builder.HasOne<Project>()
                 .WithMany()
                 .HasForeignKey(p => p.DefaultProjectId)
