+++ EasyFinance.Server/Controllers/ProjectsController.cs
@@ -1,4 +1,6 @@
-﻿using EasyFinance.Application.DTOs.AccessControl;
+﻿using System.Net;
+using System.Security.Claims;
+using EasyFinance.Application.DTOs.AccessControl;
 using EasyFinance.Application.DTOs.FinancialProject;
 using EasyFinance.Application.Features.AccessControlService;
 using EasyFinance.Application.Features.CategoryService;
@@ -6,16 +8,11 @@ using EasyFinance.Application.Features.IncomeService;
 using EasyFinance.Application.Features.ProjectService;
 using EasyFinance.Application.Mappers;
 using EasyFinance.Domain.AccessControl;
-using EasyFinance.Domain.Financial;
-using EasyFinance.Domain.FinancialProject;
 using EasyFinance.Infrastructure;
 using EasyFinance.Infrastructure.DTOs;
 using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Identity.UI.Services;
 using Microsoft.AspNetCore.JsonPatch;
 using Microsoft.AspNetCore.Mvc;
-using System.Net;
-using System.Security.Claims;
 
 namespace EasyFinance.Server.Controllers
 {
